NLB-6212: enable R33 directives
